pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app/"]

[project]
name = "travelio"
version = "0.1.0"
description = "A travel itinerary application"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.9", # For PostgreSQL
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "python-jose>=3.3.0", # For JWT tokens
    "passlib>=1.7.4", # For password hashing
    "python-multipart>=0.0.6", # For form data processing
    "scikit-learn>=1.3.2", # For clustering algorithms
    "geopy>=2.4.1", # For geospatial calculations
    "httpx>=0.25.0", # For HTTP requests
    "python-dotenv>=1.1.0",
    "google-generativeai>=0.8.5",
    "mcp-use>=1.2.8",
    "fastembed>=0.6.1",
    "google-genai>=1.12.1",
    "langchain-openai>=0.3.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.10.1",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.1",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
